Move addCommand utility from GuestUtil to BookingUtil
