+++ src/test/java/seedu/address/logic/parser/ConciergeParserTest.java
@@ -41,6 +41,7 @@ import seedu.address.model.guest.GuestNameContainsKeywordsPredicate;
 import seedu.address.model.room.Room;
 import seedu.address.model.room.RoomNumber;
 import seedu.address.model.room.booking.BookingPeriod;
+import seedu.address.testutil.BookingUtil;
 import seedu.address.testutil.EditGuestDescriptorBuilder;
 import seedu.address.testutil.GuestBuilder;
 import seedu.address.testutil.GuestUtil;
@@ -61,7 +62,7 @@ public class ConciergeParserTest {
         BookingPeriod bookingPeriod = TODAY_NEXTWEEK;
 
         AddCommand command = (AddCommand) parser.parseCommand(
-                GuestUtil.getAddCommand(guest, roomNumber, bookingPeriod));
+                BookingUtil.getAddCommand(guest, roomNumber, bookingPeriod));
         assertEquals(new AddCommand(guest, roomNumber, bookingPeriod),
                 command);
     }
